Refactor into dag with helper to simply views
