+++ airflow/www/templates/airflow/dag.html
@@ -133,9 +133,9 @@
           Next Run: <time datetime="{{ dag_model.next_dagrun }}">{{ dag_model.next_dagrun }}</time>
         </p>
       {% endif %}
-      {% if dag_model is defined and dataset_triggered_next_run_info %}
+      {% if dag_model is defined and dag_model.schedule_interval is defined and dag_model.schedule_interval == 'Dataset' %}
         <p class="label label-default" style="margin-left: 5px">
-          Next Run: {{ dataset_triggered_next_run_info }}
+          Next Run: {{ dag_model.get_dataset_triggered_next_run_info() }}
         </p>
       {% endif %}
     </h4>
